+++ heat/core/tests/test_indexing.py
@@ -1,8 +1,4 @@
-import unittest
-import os
 import heat as ht
-import torch
-
 from .test_suites.basic_test import TestCase